Fixed readByte() inconsistencies
swapped from using "int*" to a "char*", since "char*" uses 1 byte, and is the general return for "readBytes". I instead decided to just cast the "char*", to a "int*", so that data can be handled more appropriately.

"readBytes" had a minor name inconsistency of using "Offset" instead of "offset"

"readBytes" passable values were swapped as a way to be more intuitive to use.